Create Google Authentication

+++ controllers/api/v1/posts_api.js
@@ -2,7 +2,7 @@ const Post = require('../../../models/post');
 const Comment = require('../../../models/comment');
 
 
-module.exports.index = async function(req, res){
+module.exports.index = async function (req, res) {
 
     let posts = await Post.find({})
         .populate('user')
@@ -20,22 +20,28 @@ module.exports.index = async function(req, res){
 }
 
 
-module.exports.destroy = async function(req, res){
-   try {
+module.exports.destroy = async function (req, res) {
+    try {
 
-    let post = await Post.findById(req.params.id);
+        let post = await Post.findById(req.params.id);
 
-    post.remove();
+        if (post.user == req.user.id) {
+            post.remove();
 
-    await Comment.deleteMany({post : req.params.id});
+            await Comment.deleteMany({ post: req.params.id });
 
-    return res.json(200, {
-        message : "Post and associated comments deleted successfully"
-    });
+            return res.json(200, {
+                message: "Post and associated comments deleted successfully"
+            });
+        } else {
+            return res.json(401, {
+                message: 'You can not delete this post!'
+            });
+        }
 
-   } catch (error) {
-     return res.json(500, {
-        message : "Internal Server Error"
-    });
-   }
+    } catch (error) {
+        return res.json(500, {
+            message: "Internal Server Error"
+        });
+    }
 }
